+++ doc/manage_kernels_ext_npu.md
